Address review feedback from masnwilliams
- Clamp Low watermark to High+1 instead of cascading reset to defaults
- Collapse duplicate hold-clearing branches in reconcile
- Add lastApplied pruning for VMs no longer in candidate list
- Wrap sampler error with context for better caller diagnostics
- Fix err shadowing in ResolveProcessPID (socket_pid_linux.go)
- Use exact arg match in pidByCmdline to avoid substring false positives
- Extract assertActiveBallooningLifecycle to shared test helpers
