+++ lib/instances/guestmemory_active_ballooning_test_helpers_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/kernel/hypeman/lib/guestmemory"
 	"github.com/kernel/hypeman/lib/hypervisor"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -128,4 +129,31 @@ func requireManualReclaimCleared(t *testing.T, ctx context.Context, controller g
 	require.NoError(t, err)
 	requireRuntimeGuestMemoryTargetEventually(t, ctx, inst, inst.Size+inst.HotplugSize)
 	return resp
+}
+
+func assertActiveBallooningLifecycle(t *testing.T, ctx context.Context, inst *Instance) {
+	t.Helper()
+
+	assigned := inst.Size + inst.HotplugSize
+	initialTarget := requireRuntimeGuestMemoryTarget(t, ctx, inst)
+	assert.Equal(t, assigned, initialTarget, "runtime balloon target should start at full assigned memory")
+
+	controller := newActiveBallooningTestController(t, inst)
+
+	reclaimResp := requireManualReclaimApplied(t, ctx, controller, inst, 1*1024*1024*1024, 5*time.Minute)
+	require.Len(t, reclaimResp.Actions, 1)
+	assert.NotNil(t, reclaimResp.HoldUntil)
+	assert.Equal(t, int64(1*1024*1024*1024), reclaimResp.Actions[0].AppliedReclaimBytes)
+	assert.Equal(t, assigned-int64(1*1024*1024*1024), reclaimResp.Actions[0].TargetGuestMemoryBytes)
+
+	clearResp := requireManualReclaimCleared(t, ctx, controller, inst)
+	assert.Nil(t, clearResp.HoldUntil)
+
+	floorResp := requireManualReclaimApplied(t, ctx, controller, inst, assigned, 5*time.Minute)
+	require.Len(t, floorResp.Actions, 1)
+	expectedFloor := assigned / 2
+	assert.Equal(t, expectedFloor, floorResp.Actions[0].TargetGuestMemoryBytes)
+	assert.Equal(t, assigned-expectedFloor, floorResp.Actions[0].AppliedReclaimBytes)
+
+	requireManualReclaimCleared(t, ctx, controller, inst)
 }
